feat(imfrastructure): completed config for employees module

+++ Src/Domain/DevLife.Domain/Modules/Employees/EmployeeSkillModificator.cs
@@ -7,7 +7,7 @@ public class EmployeeSkillModificator : AuditableBaseEntity
 {
     public required string Name { get; set; }
     public required string Description { get; set; }
-    public required int Modificator { get; set; }
+    public required decimal Modificator { get; set; }
 
     // Relation
     public ICollection<EmployeeSkill>? EmployeeSkills {get; set;}
